refactor(anyharness): re-cure prompt.rs and runtime/tests.rs ratchets by relocation (Lane H, lead ruling)
Lead ruling on the five ratchet bumps: app/mod.rs +14, fork.rs +13,
error.rs +7 authorized (justifications in the PR body); prompt.rs and
runtime/tests.rs re-cured move-only. The in-file dispatch-classification
test module moves verbatim out of prompt.rs (584 lines now, under the
global cap, entry deleted); the same-payload-resume rung-2 test moves
verbatim into checkpoint_linkage_tests.rs beside the NULL-linkage pins it
carries (tests.rs 1218, entry tightened per the shrink-only discipline).
Helpers shared via pub(super), not copied. Moved suites rerun green by
narrow filter; check_max_lines passes.
