+++ anyharness/crates/anyharness-lib/src/domains/sessions/runtime/prompt.rs
@@ -481,7 +481,7 @@ pub(crate) enum TextPromptDispatchError {
 /// was enqueued to the actor's mailbox, not that the actor processed it, and
 /// the reply was lost either way; `ActorUnavailable` (command never enqueued)
 /// and an explicit rejection are safe to report as failed dispatch.
-fn classify_text_prompt_command_error(
+pub(super) fn classify_text_prompt_command_error(
     error: LiveSessionCommandError<PromptAcceptError>,
 ) -> TextPromptDispatchError {
     match error {
@@ -513,7 +513,7 @@ fn map_lifecycle_error_to_prompt(error: SessionLifecycleError) -> SendPromptErro
     }
 }
 
-fn durable_prompt_start_failure_code(error: &StartSessionError) -> &'static str {
+pub(super) fn durable_prompt_start_failure_code(error: &StartSessionError) -> &'static str {
     match error {
         StartSessionError::WorkspaceNotFound => "workspace_not_found",
         StartSessionError::WorkspaceDirectoryMissing { .. } => "workspace_directory_missing",
@@ -580,78 +580,5 @@ fn map_start_error_to_prompt(error: StartSessionError) -> SendPromptError {
         StartSessionError::Internal(error) | StartSessionError::AcpStart(error) => {
             SendPromptError::Internal(error)
         }
-    }
-}
-
-#[cfg(test)]
-mod dispatch_classification_tests {
-    use super::*;
-    use crate::domains::agents::model::ResolvedAgentStatus;
-
-    #[test]
-    fn send_message_start_failure_codes_do_not_expose_details() {
-        let cases = [
-            (
-                StartSessionError::WorkspaceDirectoryMissing {
-                    path: "/secret/workspace".into(),
-                },
-                "workspace_directory_missing",
-            ),
-            (
-                StartSessionError::AgentNotReady {
-                    agent_kind: "secret-agent".into(),
-                    status: ResolvedAgentStatus::Error,
-                    detail: Some("secret readiness detail".into()),
-                },
-                "agent_not_ready",
-            ),
-            (
-                StartSessionError::Internal(anyhow::anyhow!("secret")),
-                "internal",
-            ),
-            (
-                StartSessionError::AcpStart(anyhow::anyhow!("secret")),
-                "acp_start",
-            ),
-        ];
-
-        for (error, expected) in cases {
-            assert_eq!(durable_prompt_start_failure_code(&error), expected);
-        }
-    }
-
-    #[test]
-    fn response_dropped_is_a_lost_acknowledgement() {
-        assert!(matches!(
-            classify_text_prompt_command_error(LiveSessionCommandError::ResponseDropped),
-            TextPromptDispatchError::AcknowledgementLost
-        ));
-    }
-
-    #[test]
-    fn actor_unavailable_and_rejection_are_failed_dispatch() {
-        assert!(matches!(
-            classify_text_prompt_command_error(LiveSessionCommandError::ActorUnavailable),
-            TextPromptDispatchError::Dispatch(SendPromptError::Internal(_))
-        ));
-        assert!(matches!(
-            classify_text_prompt_command_error(LiveSessionCommandError::Rejected(
-                PromptAcceptError::EnqueueFailed("queue closed".to_string())
-            )),
-            TextPromptDispatchError::Dispatch(SendPromptError::Internal(_))
-        ));
-        assert!(matches!(
-            classify_text_prompt_command_error(LiveSessionCommandError::Rejected(
-                PromptAcceptError::ProductContextUnavailable {
-                    incident_id: "incident-1".to_string(),
-                    error: crate::live::sessions::product_context::AgentProductContextResolutionError::new(
-                        anyhow::anyhow!("private")
-                    ),
-                }
-            )),
-            TextPromptDispatchError::Dispatch(
-                SendPromptError::ProductContextUnavailable { incident_id, .. }
-            ) if incident_id == "incident-1"
-        ));
     }
 }
